project/requirements.txt
numpy
scikit-learn
pillow
fastapi
uvicorn
joblib
python-multipart
torch
torchvision
